Nombre del proyecto: Conciliación automática. Descripción(ocontreras): Refactorización del código.

+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/LayoutHeadersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import mx.com.nmp.pagos.mimonte.model.conciliacion.LayoutHeader;
 
 @Repository
-public interface LayoutHeadersRepository extends JpaRepository<LayoutHeader, Integer>{
+public interface LayoutHeadersRepository extends JpaRepository<LayoutHeader, Long>{
 
 }
